Conflicts:
	web/src/main/webapp/resources/assets/i18n/welcome-ukr.json

+++ service/src/main/java/com/softserve/edu/service/admin/AgreementService.java
@@ -14,15 +14,15 @@ public interface AgreementService {
 
     Agreement add(Agreement agreement);
 
-    Agreement add(Organization customer, Organization executor, String number, Long deviceCount, Date date, Device.DeviceType deviceType);
+    Agreement add(Organization customer, Organization executor, String number, int deviceCount, Date date, Device.DeviceType deviceType);
 
-    Agreement add(Long customerId, Long executorId, String number, Long deviceCount, Date date, Device.DeviceType deviceType);
+    Agreement add(Long customerId, Long executorId, String number, int deviceCount, Date date, Device.DeviceType deviceType);
 
     Agreement findAgreementById(Long id);
 
     Set<Agreement> findAll();
 
-    void update(Long agreementId, Long customerId, Long executorId, String number, Long deviceCount, Date date, Device.DeviceType deviceType);
+    void update(Long agreementId, Long customerId, Long executorId, String number, int deviceCount, Date date, Device.DeviceType deviceType);
 
     ListToPageTransformer<Agreement> getCategoryDevicesBySearchAndPagination(int pageNumber, int itemsPerPage, Map<String,
             String> searchKeys, String sortCriteria, String sortOrder);
